\name{parsec-package}
\alias{parsec-package}
\alias{parsec}
\docType{package}
\title{
\packageTitle{parsec}
}
\description{
\packageDescription{parsec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{parsec}
\packageIndices{parsec}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{parsec}

Maintainer: \packageMaintainer{parsec}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
